refactor: Remove extra debug scene and integrate the demo into the `interface` scene.
