+++ src/rust/ide/view/src/debug_scenes.rs
@@ -1,5 +1,4 @@
 //! A set of debug scenes
 
 pub mod interface;
-pub mod vcs;
 pub mod visualization;
